requirements.txt
discord.py
thmapi